+++ test/Examples.hs
@@ -51,10 +51,15 @@ main = hspec $ do
       expectFailure (\m1 m2 ->
             length (svgViaTex (m1::Exp.KripkeModelS5,0::Exp.World))
         === length (svgViaTex (m2::Exp.KripkeModelS5,0::Exp.World)) )
-  describe "SMCDEL.Symbolic.S5" $
+  describe "SMCDEL.Symbolic.S5" $ do
     prop "boolEvalViaBdd agrees on simplified formulas" $
       \(BF bf) props -> let truths = nub props in
         boolEvalViaBdd truths bf === boolEvalViaBdd truths (simplify bf)
+    prop "optimize preserves truth" $
+      \kns f -> let scene = (kns :: KnowStruct, head $ statesOf kns)
+                in isTrue scene f === isTrue (optimize defaultVocabulary scene) f
+    prop "optimize can reduce the vocabulary" $
+      expectFailure (\kns -> length (vocabOf (kns :: KnowStruct)) == length (vocabOf (optimize defaultVocabulary kns)))
   describe "SMCDEL.Other.BDD2Form" $ do
     prop "boolBddOf . formOf == id" $
       \b -> b === boolBddOf (formOf b)